basic-agents/requirements.txt
google-adk[a2a]
google-genai
anthropic